chore: Add admin guard and provider in auth configuration
This commit adds a new guard and provider for admins in the auth configuration file. The admin guard uses the JWT driver and the admins provider uses the eloquent driver. This allows for separate authentication and password reset options for admin users. The necessary changes have been made in the 'auth.php' file.

Deleted migration file for 'admin' table

This commit deletes the migration file for the 'admin' table. The migration file was named '2024_09_23_040258_create_admin_table.php' and has been removed from the project.

Add AdminAuthController and related files

This commit adds the 'AdminAuthController' and related files to the project. The 'AdminAuthController' handles the registration, login, and logout functionality for admin users. The controller is located in the 'app/Http/Controllers' directory.

Add Admin model and migration file

This commit adds the 'Admin' model and migration file to the project. The 'Admin' model extends the 'Authenticatable' class and implements the 'JWTSubject' interface. The migration file creates the 'admins' table in the database.

Create migration file for 'admins' table

This commit creates the migration file for the 'admins' table. The migration file is named '2024_09_25_114325_create_admins_table.php' and is located in the 'database/migrations' directory.